+++ programs/core/src/libraries/liquidity_math.rs
@@ -1,4 +1,4 @@
-/// Math library for liquidity
+///! Math library for liquidity
 
 use anchor_lang::require;
 use crate::error::ErrorCode;
@@ -8,16 +8,15 @@ use crate::error::ErrorCode;
 /// # Arguments
 ///
 /// * `x` - The liquidity (L) before change
-///
 /// * `y` - The delta (ΔL) by which liquidity should be changed
 ///
-pub fn add_delta(x: u32, y: i32) -> Result<u32, ErrorCode> {
-    let z: u32;
+pub fn add_delta(x: u64, y: i64) -> Result<u64, ErrorCode> {
+    let z: u64;
     if y < 0 {
-        z = x - (y.abs() as u32);
+        z = x - (y.abs() as u64);
         require!(z < x, ErrorCode::LS);
     } else {
-        z = x + (y as u32);
+        z = x + (y as u64);
         require!(z >= x, ErrorCode::LA);
     }
 
@@ -30,27 +29,27 @@ mod tests {
 
     #[test]
     fn positive_liquidity_delta() {
-        let x: u32 = 1;
-        let y: i32 = 2;
+        let x: u64 = 1;
+        let y: i64 = 2;
         assert_eq!(add_delta(x, y).unwrap(), 3);
     }
 
     #[test]
     fn negative_liquidity_delta() {
-        let x: u32 = 2;
-        let y: i32 = -1;
+        let x: u64 = 2;
+        let y: i64 = -1;
         assert_eq!(add_delta(x, y).unwrap(), 1);
     }
 
     #[test]
     #[should_panic]
     fn positive_liquidity_delta_overflow() {
-        add_delta(u32::MAX, 1).unwrap();
+        add_delta(u64::MAX, 1).unwrap();
     }
 
     #[test]
     #[should_panic]
     fn negative_liquidity_delta_underflow() {
-        add_delta(u32::MIN, -1).unwrap();
+        add_delta(u64::MIN, -1).unwrap();
     }
 }
