+++ programs/non-fungible-position-manager/src/states/non_fungible_position.rs
@@ -4,7 +4,7 @@ use anchor_lang::prelude::*;
 #[derive(Default)]
 pub struct NonFungiblePositionState {
     pub bump: u8,
-    pub liquidity: u32,
+    pub liquidity: u64,
     pub fee_growth_inside_0_x32: u64,
     pub fee_growth_inside_1_x32: u64,
     pub tokens_owed_0: u64,
